pyproject.toml
[tool.poetry]
name = "dataloader"
version = "0.1"
description = "test poc"
authors = ["shyam <shyam1011@gmail.com>"]
packages = [{include = "dataloader", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
pandas = ">2.2.0"
databricks-sql-connector = ">2.9.0"
pyspark = "*"

[tool.poetry.scripts]
dataloader = "dataloader.load_data:load_data"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
